email_tracker/requirements.txt
asgiref==3.8.1
dj-database-url==2.2.0
Django==5.0.6
django-ckeditor==6.7.1
django-js-asset==2.2.0
gunicorn==22.0.0
packaging==24.1
python-decouple==3.8
sqlparse==0.5.0
typing_extensions==4.12.2
